Database file
This is the database exported to a .sql file, so I can work on this project from another pc.

This project is of right now in a hold, because I have a lot of work at my job and a lot of work in school.
When I have time, I will finish this project and then I will make a Raspberry Pi installed on a, at least 2TB harddisk, so it can hold all of these series and eventually movies. It will then be accessible and you will be able to watch my hosted series and movies for free without any ads. The Login form will be updated before release so it has hashed and salted usernames and passwords. And then It will have a limit of 3 tries an hour. If you wrongly input a username and/or password 3 times, you will be locked out at least an hour from further tries. This will prevent bruteforcing your way in.