Video player almost working
I got it to select the correct episode when watching a series. Now I am working on making it not exit fullscreen when a video has ended and it selects the next one. Before it did exit, because it reloaded the page with the new episode. I will try and avoid that by just changing the source of the video when it has ended.
But it is a harder task then anticipated, because there are a lot of stuff to avoid when increasing the episode number and especially the season. But I almost got it working.