It now plays next episode
It now plays next episode by changing source of video. With this, it does not go out of fullscreen.

Next step is to highlight currentlæy played episode in the table underneath and change the title above video player.

Next step again is to make season tabs, so the user can easily change season and change episode from another season.