Admin Panel + Add episodes with text file
I have made the Admin Panel work again. So you can now update the thumbnails for seasons. Add series and add episodes.
While I was adding some episodes to one of the series on there, I thought that it was a bit to time consuming. So I made started testing something. You can now add a lot of episodes at one, with a text file. You just need to follow the intructions in the text file, and it will be good.
So now it is possible to add many episodes from different seasons and series.
I will make it so you can write it in a text file on the pc and then upload the file and add them through there.
Right now it just takes the testfile.txt and reads the data in there.